Split main async execution to different class
